Moving signature creation logic to a seperate class
